Sprint 29 -
Lista de Atividades:
Atendido -  Reduzir a espessura da linha que separa as atividades da lista de atividades.
Atendido (Aumentada a altura da linha) - Aumentar um pouco a fonte das letras das atividade da lista de atividades. Senão for possivel, aumentar a altura da linha contendo a atividade. Isso deixará as informções mais legíveis e atividade mais fácil de selecionar.

Atividade:
Atendido - Caso finalize a atividade com um valor de area realizada diferente da area do talhão, abrir dialog para justificar o motivo e gravar nas observações. Ex: Area tem 48ha e a pessoa vai finalizar com area realizada de 45ha.
Atendido - Ao editar atividade já finalizada, liberar acesso a todos os botões da atividade. Ex: Ao editar, apresenta a regra de atividade nova habilitando somente Calibração. Ao editar a atividade que finalizou, os 3 botões devem estar habilitados.

Calibração:
Atendido -   Mudar a mensagem quando sai da calibração sem conclui-la. “Caso clique em sim, a calibração em andamento será descartada.”

Apontamento Diário:
Atendido (Colocada uma borda em volta de cada View)-  Colocar uma linha separando rendimento de insumos.
Atendido (Colocada uma caixa de texto ensinando a como adicionar a aplicação do insumo) - Em insumos, colocar um inputtext ou algum simbolo que indique o campo da grid para adicionar a quantidade. Sem nada fica sem saber o que fazer.
Atendido - Quando apresentar o erro de que a area inserida passou, informar o valor da diferença dos hectares. Ex: Area Talhao=45ha, Area Realizada 48ha, Exibir que passou 3ha.
Corrigido - Percentual da diferença entre qtde programada com relação a aplicada está com erro. Ex: Considerando que o planejado para o Talhão de área de 48 ha, para Polyblen é de 8640 e o aplicado nesse caso foi de 8856, daria uma diferença de 2,49% (216 a mais do indicado). Porém na tela fala que está 4820%.

Avaliação Qualidade:
Não foi necessário corrigir, pois não havia erro- Verificar as regras da qualidade. A profundiade por exemplo está com erro. Se vc coloca 100, ele não considera “Não Conforme” sendo que deveria por estar fora dos 5%. Foi percebido falha só nesse item, mas favor verificar todas regras para garantirmos o correto funcionamento.
Corrigido(Adicionados métodos getLatitude() e getLongitude()) - Não encontrei no código o import da biblioteca location nem a função getLatitude() ou getLongitude(). Como está pegando as coordendadas do ponto? Caso não esteja pegando, implementar a coleta das coordenadas gps do dispositivo no momento de registro do ponto.

+++ app/src/main/java/com/example/forestsys/classes/AVAL_PONTO_SUBSOLAGEM.java
@@ -43,11 +43,11 @@ public class AVAL_PONTO_SUBSOLAGEM {
 
     private double VALOR_INDICADOR;
 
-    private int COORDENADA_X;
+    private double COORDENADA_X;
 
-    private int COORDENADA_Y;
+    private double COORDENADA_Y;
 
-    public AVAL_PONTO_SUBSOLAGEM(int ID_PROGRAMACAO_ATIVIDADE, @NonNull String DATA, int PONTO, int ID_ATIVIDADE, int ID_INDICADOR, double VALOR_INDICADOR, int COORDENADA_X, int COORDENADA_Y) {
+    public AVAL_PONTO_SUBSOLAGEM(int ID_PROGRAMACAO_ATIVIDADE, @NonNull String DATA, int PONTO, int ID_ATIVIDADE, int ID_INDICADOR, double VALOR_INDICADOR, double COORDENADA_X, double COORDENADA_Y) {
         this.ID_PROGRAMACAO_ATIVIDADE = ID_PROGRAMACAO_ATIVIDADE;
         this.DATA = DATA;
         this.PONTO = PONTO;
@@ -107,19 +107,19 @@ public class AVAL_PONTO_SUBSOLAGEM {
         this.VALOR_INDICADOR = VALOR_INDICADOR;
     }
 
-    public int getCOORDENADA_X() {
+    public double getCOORDENADA_X() {
         return COORDENADA_X;
     }
 
-    public void setCOORDENADA_X(int COORDENADA_X) {
+    public void setCOORDENADA_X(double COORDENADA_X) {
         this.COORDENADA_X = COORDENADA_X;
     }
 
-    public int getCOORDENADA_Y() {
+    public double getCOORDENADA_Y() {
         return COORDENADA_Y;
     }
 
-    public void setCOORDENADA_Y(int COORDENADA_Y) {
+    public void setCOORDENADA_Y(double COORDENADA_Y) {
         this.COORDENADA_Y = COORDENADA_Y;
     }
 }
